+++ dcrjson/walletsvrresults.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2014 The btcsuite developers
-// Copyright (c) 2015 The Decred developers
+// Copyright (c) 2015-2016 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -178,5 +178,5 @@ type ValidateAddressWalletResult struct {
 // GetBestBlockResult models the data from the getbestblock command.
 type GetBestBlockResult struct {
 	Hash   string `json:"hash"`
-	Height int32  `json:"height"`
+	Height int64  `json:"height"`
 }
